python_server/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
python-multipart==0.0.6
mysql-connector-python==8.2.0
pyodbc==5.0.1
pyjwt==2.8.0
pydantic==2.4.2
python-jose==3.3.0
bcrypt==4.0.1
httpx==0.25.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.10.1
isort==5.12.0
flake8==6.1.0 